requirements.txt
accelerate==1.10.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colpali_engine @ git+https://github.com/illuin-tech/colpali@9bee9b2b72e37cde453d3e4656a1f032bad9194e
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
peft==0.16.0
pillow==11.3.0
psutil==7.0.0
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scipy==1.16.1
selenium==4.35.0
setuptools==80.9.0
six==1.17.0
smolagents==1.21.3
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.14.0
tokenizers==0.21.4
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
